+++ src/PBRTSharp/Core/Vectors/Vector3i.cs
@@ -21,8 +21,8 @@ namespace PBRTSharp.Core.Vectors
         // Operator overloads
         public static Vector3i operator +(in Vector3i v1, in Vector3i v2) => new Vector3i(v1.X + v2.X, v1.Y + v2.Y, v1.Z + v2.Z);
         public static Vector3i operator -(in Vector3i v1, in Vector3i v2) => new Vector3i(v1.X - v2.X, v1.Y - v2.Y, v1.Z - v2.Z);
-        public static Vector3f operator *(in double d, in Vector3i v) => new Vector3f(d * v.X, d * v.Y, d * v.Z);
-        public static Vector3f operator /(in Vector3i v, in double d) => 1.0d / d * v;
+        public static Vector3i operator *(in double d, in Vector3i v) => new Vector3i((int)(d * v.X), (int)(d * v.Y), (int)(d * v.Z));
+        public static Vector3i operator /(in Vector3i v, in double d) => 1.0d / d * v;
         public static Vector3i operator -(in Vector3i v) => new Vector3i(-v.X, -v.Y, -v.Z);
         public static bool operator ==(Vector3i v1, Vector3i v2) => v1.Equals(v2);
         public static bool operator !=(Vector3i v1, Vector3i v2) => !(v1 == v2);
@@ -50,7 +50,7 @@ namespace PBRTSharp.Core.Vectors
         }
         public double LengthSquared() => (X * X) + (Y * Y) + (Z * Z);
         public double Length() => Math.Sqrt(LengthSquared());
-        public Vector3f Normalize() => this / Length();
+        public Vector3i Normalize() => this / Length();
         public double MinComponent() => Math.Min(X, Math.Min(Y, Z));
         public double MaxComponent() => Math.Max(X, Math.Max(Y, Z));
         public double MaxDimension() => X > Y ? (X > Z ? 0 : 2) : (Y > Z ? 1 : 2);
